Fix GB2Byte with int64

+++ utils/calc.go
@@ -1,5 +1,5 @@
 package utils
 
-func GB2Byte(GB int) int {
-	return GB * 1024 * 1024 * 1024
+func GB2Byte(GB int) int64 {
+	return int64(GB * 1024 * 1024 * 1024)
 }
